requirements.txt
Django==4.0.7
djangorestframework==3.13.1
requests~=2.27.1
python-decouple==3.6
whitenoise==6.0.0
dj-database-url==0.5.0
django-cors-headers==3.11.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
sentry-sdk==1.9.10
numpy==1.22.3
reportlab == 3.6.9
djlint==1.0.3
django-storages==1.12.3
boto3==1.23.1
django-cleanup==6.0.0
pre-commit==2.19.0
scrapy==2.6.3
coverage==6.4.4
uvicorn==0.18.3